requirements.txt
Django>=4.1,<4.2
djangorestframework>=3.14.0,<3.15
drf-spectacular>=0.26,<0.27
psycopg2>=2.9,<3.0